Fixes to remove compiler warnings
Implement fixes for all warnings reported by compiler (GCC 4.3.4 with
-W -Wall flags). These fall into three main categories

- Local scope unused variables, fix by deletion or ignoring return
parameters.

- Unused function arguments, fix by commenting out argument name. This
helps to retain the context of the argument.

- Wrong type comparison (e.g. int -> size_t), fix by using correct type
or cast where behaviour uncertain.

+++ source/bxgeomtools/include/geomtools/polycone.h
@@ -96,7 +96,7 @@ namespace geomtools {
 
     unsigned int number_of_frustra() const;
 
-    void get_frustrum(int i_, frustrum_data &) const;
+    void get_frustrum(size_t i_, frustrum_data &) const;
 
     const rz_col_type & points () const;
